Add a way to Remove User from a Role (Issue #14632)
