+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Items/RemoveUserRoleTask.Fields.Design.cshtml
@@ -1,0 +1,9 @@
+@using OrchardCore.Workflows.ViewModels
+@using OrchardCore.Workflows.Helpers
+@using OrchardCore.Roles.Workflows.Activities
+@using OrchardCore.Roles.Workflows.ViewModels
+@model ActivityViewModel<RemoveUserRoleTask>
+<header>
+    <h4><i class="fa-solid fa-user" aria-hidden="true"></i>@Model.Activity.GetTitleOrDefault(() => T["Remove user from role"])</h4>
+</header>
+<em>@T["{0} from role {1}", Model.Activity.UserName, Model.Activity.RoleName]</em>
