+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Items/RemoveUserRoleTask.Fields.Edit.cshtml
@@ -1,4 +1,4 @@
-@using OrchardCore.Users.Workflows.ViewModels
+@using OrchardCore.Roles.Workflows.ViewModels
 @model RemoveUserRoleTaskViewModel
 
 <div class="mb-3" asp-validation-class-for="UserName">
