+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Shared/Components/SelectRoles/Default.cshtml
@@ -1,10 +1,10 @@
 @model SelectRolesViewModel
 
-<div class="d-flex flex-wrap">
+<div class="mb-3">
     @for (var i = 0; i < Model.RoleSelections.Count; i++)
     {
         var role = Model.RoleSelections[i];
-        <div class="checkbox pr-2">
+        <div class="form-check">
             <label for="@Html.IdFor(m => m.RoleSelections[i])">
                 <input id="@Html.IdFor(m => m.RoleSelections[i])" name="@Model.HtmlName" type="checkbox" checked="@role.IsSelected" value="@role.Item" />
                 @role.Item
