+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Workflows/Activities/SelectUsersInRoleTask.cs
@@ -1,46 +1,45 @@
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Localization;
+using OrchardCore.Users;
 using OrchardCore.Users.Models;
-using OrchardCore.Users.Services;
 using OrchardCore.Workflows.Abstractions.Models;
 using OrchardCore.Workflows.Activities;
 using OrchardCore.Workflows.Models;
 using OrchardCore.Workflows.Services;
 
-namespace OrchardCore.Users.Workflows.Activities
+namespace OrchardCore.Roles.Workflows.Activities
 {
-    public class RemoveUserRoleTask : TaskActivity
+    public class SelectUsersInRoleTask : TaskActivity
     {
         private readonly UserManager<IUser> _userManager;
-        private readonly IUserService _userService;
         private readonly IWorkflowExpressionEvaluator _expressionEvaluator;
         protected readonly IStringLocalizer S;
 
-        public RemoveUserRoleTask(UserManager<IUser> userManager, IUserService userService, IWorkflowExpressionEvaluator expressionvaluator, IStringLocalizer<RemoveUserRoleTask> localizer)
+        public SelectUsersInRoleTask(UserManager<IUser> userManager, IWorkflowExpressionEvaluator expressionvaluator, IStringLocalizer<SelectUsersInRoleTask> localizer)
         {
             _userManager = userManager;
-            _userService = userService;
             _expressionEvaluator = expressionvaluator;
             S = localizer;
         }
 
-        public override string Name => nameof(RemoveUserRoleTask);
+        public override string Name => nameof(SelectUsersInRoleTask);
 
-        public override LocalizedString DisplayText => S["Remove User Role Task"];
+        public override LocalizedString DisplayText => S["Select Users in Role Task"];
 
         public override LocalizedString Category => S["User"];
 
-        public WorkflowExpression<string> UserName
+        public WorkflowExpression<string> OutputKeyName
         {
             get => GetProperty(() => new WorkflowExpression<string>());
             set => SetProperty(value);
         }
 
-        public WorkflowExpression<string> RoleName
+        public IEnumerable<string> Roles
         {
-            get => GetProperty(() => new WorkflowExpression<string>());
+            get => GetProperty(() => new List<string>());
             set => SetProperty(value);
         }
 
@@ -51,24 +50,25 @@ namespace OrchardCore.Users.Workflows.Activities
 
         public override async Task<ActivityExecutionResult> ExecuteAsync(WorkflowExecutionContext workflowContext, ActivityContext activityContext)
         {
-            var userName = await _expressionEvaluator.EvaluateAsync(UserName, workflowContext, null);
-            var roleName = await _expressionEvaluator.EvaluateAsync(RoleName, workflowContext, null);
+            var propKeyName = await _expressionEvaluator.EvaluateAsync(OutputKeyName, workflowContext, null);
 
-            var user = (User)await _userService.GetUserAsync(userName);
-
-            if (user != null)
+            if (!string.IsNullOrEmpty(propKeyName))
             {
-                if (user.RoleNames.Contains(roleName))
+                if (Roles.Any())
                 {
-                    await _userManager.RemoveFromRoleAsync(user, roleName);
+                    HashSet<IUser> usersInRole = new HashSet<IUser>();
+                    foreach (var roleName in Roles)
+                    {
+                        usersInRole.UnionWith(await _userManager.GetUsersInRoleAsync(roleName));
+                    }
+                    if (usersInRole.Any())
+                    {
+                        workflowContext.Output[propKeyName] = usersInRole.Select(u => (u as User).UserId).ToArray();
+                        return Outcomes("Done");
+                    }
                 }
-
-                return Outcomes("Done");
             }
-            else
-            {
-                return Outcomes("Failed");
-            }
+            return Outcomes("Failed");
         }
     }
 }
