+++ src/OrchardCore.Modules/OrchardCore.Roles/Views/Workflows/ViewModels/SelectUsersInRoleTaskViewModel.cs
@@ -1,13 +1,13 @@
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace OrchardCore.Users.Workflows.ViewModels
+namespace OrchardCore.Roles.Workflows.ViewModels
 {
     public class SelectUsersInRoleTaskViewModel
     {
         [Required]
         public string OutputKeyName { get; set; }
 
-        [Required]
-        public string RoleName { get; set; }
+        public IEnumerable<string> Roles { get; set; }
     }
 }
