--- src/OrchardCore.Modules/OrchardCore.Users/Views/Items/SelectUsersInRoleTask.Fields.Design.cshtml
@@ -1,9 +1,0 @@
-@using OrchardCore.Workflows.ViewModels
-@using OrchardCore.Workflows.Helpers
-@using OrchardCore.Users.Workflows.Activities
-@using OrchardCore.Users.Workflows.ViewModels
-@model ActivityViewModel<SelectUsersInRoleTask>
-<header>
-    <h4><i class="fa-solid fa-user" aria-hidden="true"></i>@Model.Activity.GetTitleOrDefault(() => T["Select users in role"])</h4>
-</header>
-<em>@T["Stores users from role {0} to {1}", Model.Activity.RoleName, Model.Activity.OutputKeyName]</em>
