--- src/OrchardCore.Modules/OrchardCore.Users/Views/Items/SelectUsersInRoleTask.Fields.Edit.cshtml
@@ -1,16 +1,0 @@
-@using OrchardCore.Users.Workflows.ViewModels
-@model SelectUsersInRoleTaskViewModel
-
-<div class="mb-3" asp-validation-class-for="OutputKeyName">
-    <label asp-for="OutputKeyName">@T["OutputKeyName"]</label>
-    <input type="text" asp-for="OutputKeyName" class="form-control code" />
-    <span asp-validation-for="OutputKeyName"></span>
-    <span class="hint">@T["The Output Key Name to save list. With Liquid support."]</span>
-</div>
-
-<div class="mb-3" asp-validation-class-for="RoleName">
-    <label asp-for="RoleName">@T["RoleName"]</label>
-    <input type="text" asp-for="RoleName" class="form-control code" />
-    <span asp-validation-for="RoleName"></span>
-    <span class="hint">@T["The Role to search for. With Liquid support."]</span>
-</div>
